Begin adding some validation to the new models
This is largely borrowed from ECF at the moment

+++ db/migrate/20231208113756_create_contracts.rb
@@ -4,12 +4,12 @@ class CreateContracts < ActiveRecord::Migration[7.0]
       t.boolean :special_course
       t.references :statement, null: false, foreign_key: true
       t.references :course, null: false, foreign_key: true
-      t.decimal :recruitment_target
+      t.integer :recruitment_target
       t.decimal :per_participant
-      t.decimal :output_payment_percentage
-      t.decimal :number_of_payment_periods
-      t.decimal :service_fee_percentage
-      t.decimal :service_fee_installments
+      t.integer :output_payment_percentage
+      t.integer :number_of_payment_periods
+      t.integer :service_fee_percentage
+      t.integer :service_fee_installments
       t.references :cohort, null: false, foreign_key: true
       t.references :lead_provider, null: false, foreign_key: true
